Enhance SimpleBench with timeout handling and validation improvements
- Added validation for the 'n' value in the benchmark decorator to ensure it is a positive integer.
- Introduced a DEFAULT_TIMEOUT_GRACE_PERIOD constant for managing timeout behavior.
- Expanded error tags in exceptions to cover invalid timeout types and values.
- Updated the Iteration class to accept both int and float for the 'n' parameter, reflecting its flexibility.
- Implemented a timeout mechanism in the SimpleRunner to enforce execution limits on benchmarks.
- Refactored timeout-related exceptions to use a consistent naming convention.
- Adjusted test cases to validate new behaviors and ensure compatibility with the updated types.

+++ tests/test_decorators.py
@@ -635,7 +635,7 @@ def test_decorator_use_field_for_n_valid() -> None:
     assert test_case.results is not None, "Expected results to be collected."
     assert len(test_case.results) == 3, "Expected three results for the three size variations."
     expected_n_values: set[int] = set(sizes)
-    actual_n_values: set[int] = set(result.n for result in test_case.results)
+    actual_n_values: set[float] = set(result.n for result in test_case.results)
     assert actual_n_values == expected_n_values, "The n values in results do not match the expected sizes."
